+++ RELEASE_NOTES.md
@@ -1,10 +1,11 @@
 # Release Notes
 
-## 1.0.0-rc.2 - 2022-05-16
+## 1.0.0-rc.2 - 2022-05-25
 
-<version:1.0.0-rc.2+20220516>
+<version:1.0.0-rc.2+20220525>
 
-- #26 Bookmarks
+- #97 Model Deployments
+- #26 Bookmarks (PbixProj v0.12 schema)
 - Fixed #109 'pbi-tools info' no longer fails when another instance of SSAS runs on the same machine
 
 ## 1.0.0-rc.1 - 2022-03-06